+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ldubok <ldubok@student.42warsaw.pl>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/08/11 14:09:35 by ldubok            #+#    #+#             */
-/*   Updated: 2026/08/13 13:12:55 by ldubok           ###   ########.fr       */
+/*   Updated: 2026/08/13 13:34:32 by ldubok           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,6 @@ int	ft_putchar (char);
 int	ft_putstr (char *);
 int	ft_putnbr (int);
 int	ft_putunbr (unsigned int);
-int ft_puthexa (unsigned long, char);
+int ft_puthexa (unsigned int, char);
 int ft_putptr (void *);
